Move input-transform code in separate module
