+++ annif/transformer/langfilter.py
@@ -1,0 +1,19 @@
+# TODO Add dcostring
+import annif
+
+
+class LangFilter():
+
+    def __init__(self, *posargs):  # TODO Error on unknown params
+        pass
+
+    def transform_text(self, text):
+        # print('lang filter in action')
+        # TODO Implement functionality
+        return text
+
+    def transform_corpus(self, corpus):
+        # print('lang filter in action')
+        # TODO Implement functionality
+        return corpus
+
